backend/requirements.txt

fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.29
email-validator==2.0.0
python-dotenv==1.1.0
pydantic==1.10.12
openai==0.28
truthbrush==0.1.9
bs4==0.0.2
requests==2.32.3

